rename to ButtonIconized
